+++ ui/dev/src/pages/TestAutoComplete.vue
@@ -1,7 +1,7 @@
 <template>
   <q-page class="justify-center row" padding>
     <div class="col-12">
-      <qas-auto-complete v-model="values" class="col-12" clear-icon="o_clear" dropdown-icon="o_arrow_drop_down" multiple no-option-label="Selecione um empreendimento." :options="options">
+      <qas-select v-model="values" class="col-12" clear-icon="o_clear" dropdown-icon="o_arrow_drop_down" label="Select" multiple no-option-label="Selecione um empreendimento." :options="options">
         <template #option="scope">
           <q-item v-bind="scope.itemProps" v-on="scope.itemEvents">
             <q-item-section>
@@ -11,9 +11,11 @@
         </template>
 
         <template #selected-item="scope">
-          {{ scope.opt.label }} - {{ scope.opt.data.reference }}
+          {{ scope.opt.label }} - {{ scope.opt.data.reference }},
         </template>
-      </qas-auto-complete>
+      </qas-select>
+
+      <qas-select v-model="searchable" class="col-12 q-mt-lg" clear-icon="o_clear" dropdown-icon="o_arrow_drop_down" label="Searchable" multiple no-option-label="Selecione um empreendimento." :options="options" searchable />
 
       <qas-form-generator v-model="formGenerator.values" class="q-mt-lg" columns="fit" :fields="formGenerator.fields" :fields-props="fieldsProps" />
     </div>
@@ -24,15 +26,22 @@
 export default {
   data () {
     return {
-      values: '',
+      values: [6, 7, 4],
+      searchable: [6, 7, 4],
       options: [
-        // { label: 'Apto 1000', value: 1, data: { reference: 'Bloco 9' } },
-        // { label: 'Apto 1001', value: 2, data: { reference: 'Bloco 10' } }
+        { label: 'Apto 1000', value: 1, data: { reference: 'Bloco 9' } },
+        { label: 'Apto 1001', value: 2, data: { reference: 'Bloco 10' } },
+        { label: 'Apto 2001', value: 3, data: { reference: 'Bloco 2' } },
+        { label: 'Apto 2010', value: 4, data: { reference: 'Bloco 8' } },
+        { label: 'Apto 8000', value: 5, data: { reference: 'Bloco 3' } },
+        { label: 'Apto 5200', value: 6, data: { reference: 'Bloco 5' } },
+        { label: 'Apto 7845', value: 7, data: { reference: 'Bloco 2' } },
+        { label: 'Apto 2492', value: 8, data: { reference: 'Bloco 1' } }
       ],
 
       formGenerator: {
         values: {
-          autocomplete: ''
+          select: [6, 7, 4]
         },
 
         fields: {
@@ -42,11 +51,18 @@ export default {
             name: 'select',
             options: [
               { label: 'Apto 1000', value: 1, data: { reference: 'Bloco 9' } },
-              { label: 'Apto 1001', value: 2, data: { reference: 'Bloco 10' } }
+              { label: 'Apto 1001', value: 2, data: { reference: 'Bloco 10' } },
+              { label: 'Apto 2001', value: 3, data: { reference: 'Bloco 2' } },
+              { label: 'Apto 2010', value: 4, data: { reference: 'Bloco 8' } },
+              { label: 'Apto 8000', value: 5, data: { reference: 'Bloco 3' } },
+              { label: 'Apto 5200', value: 6, data: { reference: 'Bloco 5' } },
+              { label: 'Apto 7845', value: 7, data: { reference: 'Bloco 2' } },
+              { label: 'Apto 2492', value: 8, data: { reference: 'Bloco 1' } }
             ],
-            type: 'auto-complete',
-            // readonly: true
-            multiple: true
+            type: 'select',
+            // readonly: true,
+            multiple: true,
+            searchable: true
           }
         }
       }